@@ -9,21 +9,24 @@
 		"optimize-autoloader": true
 	},
 	"require": {
-		"php": ">=7.2",
-		"krinkle/toollabs-base": "1.0.2",
-		"krinkle/intuition": "2.0.0"
+		"php": ">=7.3",
+		"krinkle/toollabs-base": "^2.0.0",
+		"krinkle/intuition": "^2.0.0"
 	},
 	"require-dev": {
-		"jakub-onderka/php-parallel-lint": "^1.0.0",
-		"squizlabs/php_codesniffer": "^2.9.1",
-		"phpunit/phpunit": "^8.2.3"
+		"php-parallel-lint/php-parallel-lint": "1.3.2",
+		"mediawiki/mediawiki-phan-config": "0.11.1",
+		"phpunit/phpunit": "^8.5.27 || ^9",
+		"squizlabs/php_codesniffer": "^3.6"
 	},
 	"scripts": {
 		"test": [
 			"parallel-lint . --exclude vendor",
 			"phpunit",
+			"@phan",
 			"phpcs -p -s"
 		],
-		"cover": "phpunit --coverage-html coverage/"
+		"cover": "phpunit --coverage-html coverage/",
+		"phan": "phan --allow-polyfill-parser --no-progress-bar"
 	}
 }
